Merge branch 'dev' of github.com:oggioniale/ReLTER into dev

# Conflicts:
#	R/get_activity_info.R
#	R/get_dataset_info.R

+++ R/get_ilter_parameters.R
@@ -1,12 +1,10 @@
 #' Obtain information about the parameters collected of all sites ILTER.
-#' @description Return a `tibble` object containing Parameter collected by all
+#' @description Return a `tibble` object containing Parameters collected by all
 #' of the \href{https://www.ilter.network/network/global-coverage}{ILTER sites
-#' (more than 1200 around the world)},
-#' as a stored into \href{https://deims.org}{DEIMS-SDR}.
+#' (more than 1200 around the world)}, available from \href{https://deims.org}{DEIMS-SDR}.
 #' 
-#' This function aims to gather in a unique tibble all the Parameters
-#' the whole of ILTER sites. If you want to obtain this
-#' results, it should be noted that the execution time for this function
+#' This function gathers in a unique tibble all the Parameters
+#' from all ILTER sites. Note that the execution time for this function
 #' is very high.
 #' 
 #' If the objective is obtain information about Parameters
@@ -15,11 +13,10 @@
 #' \href{get_site_info.html}{`get_site_info()`}) or using
 #' other methods
 #' (\href{../../articles/sites_information.html}{How to about sites informations}).
-#' @param sitesNum A `integer`. It is the number of the sites that are
-#' read to get the information. Use this parameter only for get
-#' example of the output of this function. If the value of sitesNum is
-#' 0 (default) all the ILTER sites will be parsed and the waiting time will
-#' be long.
+#' @param sitesNum A `integer`. The number of the sites that are
+#' read to get the information. Use this parameter only to sample the output
+#' of this function. If the value of sitesNum is #' 0 (default) 
+#' all the ILTER sites will be parsed and the waiting time will be long.
 #' @return The output of the function is a `tibble` containing the list
 #' of parameters and their URI (Uniform Resource Identifier) collected
 #' in all ILTER sites.
@@ -79,7 +76,7 @@ get_ilter_parameters <- function(sitesNum = 0) {
     )
     uniqueSitesParameters
   } else {
-    message("\n----\nThe `sitesNum` value must be a double (e.g. 10, 24, etc.).
+    message("\n----\nThe `sitesNum` value must be a numeric (e.g. 10, 24, etc.).
 Please check again the value of `sitesNum`.\n----\n")
     uniqueSitesParameters <- NULL
   }
